Added a "do import" button
Moved the choose file button and added an additional button to do work underneath the "Import Project" in "Settings". Unable to get the project to load. Need to look at possibly needing to un-encode the file done in lines 67 and 72 of the main.js page. Did get them both to be hidden until "Import Project" was clicked.