Style fixes
Removed lines of style from the overlay page and moved to CSS file.
Organized CSS file so that sections match the sections of comments found in the overlay file.
Adjusted look of Settings, Start and Stop buttons so that they all have a little space between them.
Found unnecessary lines of code. Removed lines of code.
Pondered the irony between hyphenated and non-hyphenated.
Changed the hover over cursor for the dropdown menu so that it was only a cursor over the buttons instead of over the entire block.